% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{historical_measure_data}
\alias{historical_measure_data}
\title{All historical data for specific measures and hospitals}
\usage{
historical_measure_data(measure_ids = NULL, provider_ids = NULL)
}
\arguments{
\item{measure_ids}{A vector of measure_id values to include. If NULL (the
default), include all measure_ids.}

\item{provider_ids}{A vector of PROVIDER_ID values to include. If NULL (the
default), include all PROVIDER_IDs.}
}
\value{
A data.frame with Care Compare data for the given measures and
    providers, with columns report_date, PROVIDER_ID, measure_id, score.
}
\description{
All historical data for specific measures and hospitals
}
