% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_info_df}
\alias{hospital_info_df}
\title{Hospital information}
\format{
A data.frame with 5813 rows and 39 columns
\describe{
  \item{PROVIDER_ID}{The 6-digit hospital identifier from CMS}
  \item{hospital_name}{The hospital's name, in caps-lock}
  \item{hospital_type}{One of 6 hospital types}
  \item{address}{Street address}
  \item{city}{City}
  \item{state}{State}
  \item{zip}{Zip code}
  \item{county}{County}
  \item{phone}{Phone number}
  \item{star_rating_yyyyMM}{The hospital's star rating listed in the Care
    Compare report released in the month yyyy-MM}
  \item{peer_group_yyyyMM}{The peer group the hospital was in for the star
    rating listed in the Care Compare report released in the month yyyy-MM.
    If the hospital had a star rating, this is a count of the number of
    measure groups with at least 3 measures reported.}
}
}
\usage{
hospital_info_df
}
\description{
One row per hospital
}
\keyword{datasets}
