% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmeans_cpp}
\alias{kmeans_cpp}
\title{my k-means function}
\usage{
kmeans_cpp(x, centers, maxiter, strict)
}
\arguments{
\item{x}{the data to be clustered}

\item{centers}{initial centers}

\item{maxiter}{the max iter}

\item{strict}{how strict?}
}
\description{
my k-means function
}
