% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_info_df}
\alias{measure_info_df}
\title{Measure information}
\format{
A data.frame of
\describe{
  \item{report_date}{}
  \item{measure_group}{The domain the measure is grouped in: Mortality,
    Readmission, Patient Experience, Safety or Process}
  \item{measure_id}{The measure's current ID}
  \item{measure_name}{A descriptive name for the measure, taken from the Care Compare reports}
  \item{higher_is_better}{TRUE if higher score values are better, otherwise FALSE}
  \item{data_start_date}{The start date of the period the measure is reporting on}
  \item{data_end_date}{The end date of the period the measure is reporting on}
  \item{data_start_quarter}{The start quarter of the period the measure is reporting on}
  \item{data_end_quarter}{The end quarter of the period the measure is reporting on}
  \item{source_filename}{The name of the file in the Care Compare report
    that this measure's data was found in.}
}
}
\usage{
measure_info_df
}
\description{
One row per measure and Care Compare report the measure appears in
}
\keyword{datasets}
