% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sas_package}
\alias{sas_package}
\title{SAS Package input & output data}
\format{
A list of
\describe{
  \item{key}{The month that the Care Compare report with this data was released}
  \item{value}{A data.frame to be used as input to the star rating algorithm,
    containing all hospitals' measure data from the given Care Compare report}
}
}
\usage{
sas_package
}
\description{
When new star ratings are issued, CMS posts the algorithm in the SAS Package
here:
\url{https://qualitynet.cms.gov/inpatient/public-reporting/overall-ratings/sas}
}
\details{
This list contains the input and output data for the April 2021 and July 2022
releases of the SAS package.
}
\keyword{datasets}
