% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans.R
\name{sas_proc_fastclus}
\alias{sas_proc_fastclus}
\title{Approximate translation of the SAS procedure "fastclus" to R}
\usage{
sas_proc_fastclus(x, centers, iter.max = 10, strict = Inf)
}
\value{
the means
}
\description{
Approximate translation of the SAS procedure "fastclus" to R
}
