% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_computation.R
\name{sas_step3_stars}
\alias{sas_step3_stars}
\title{Step 3 of the star rating algorithm - use clustering to compute stars}
\usage{
sas_step3_stars(group_scores_dt)
}
\arguments{
\item{group_scores_dt}{a data.table}
}
\value{
a data.table
}
\description{
Step 3 of the star rating algorithm - use clustering to compute stars
}
