% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{star_algorithm_input_df}
\alias{star_algorithm_input_df}
\title{Star Rating Algorithm Input Data}
\format{
A data.frame with 98,055 rows and 96 columns
\describe{
  \item{report_date}{The first of the month that the Care Compare report
    with this data was released}
  \item{PROVIDER_ID}{The 6-digit hospital identifier from CMS}
  \item{94 measure columns}{The hospital's measure scores and denominators from
    the given report_date}
}
}
\usage{
star_algorithm_input_df
}
\description{
A data.frame that's a union of SAS Package input tables for each Care Compare
report posted by CMS since January 2018.
}
\keyword{datasets}
