% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{star_rating_input}
\alias{star_rating_input}
\title{Get star rating algorithm input data from a specific Care Compare report}
\usage{
star_rating_input(report_date)
}
\arguments{
\item{report_date}{The first day of the month that a Care Compare report
was released}
}
\value{
a data.frame of data that can be used as input to ``
}
\description{
Get star rating algorithm input data from a specific Care Compare report
}
