% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_computation.R
\name{sas_step1_standardize}
\alias{sas_step1_standardize}
\title{Step 1 of the star rating algorithm - standardize measure scores}
\usage{
sas_step1_standardize(sas_input_dt, star_version = "202207")
}
\arguments{
\item{sas_input_dt}{Input data, same format as the 2021-04 SAS example input}

\item{star_version}{Either "202207" (the default) or "202104"}
}
\value{
a data.table
}
\description{
Step 1 of the star rating algorithm - standardize measure scores
}
